// frontend/src/models/expert/types/CourseTranslate.tsx
import type {Language} from "../../types/Language.tsx";
import type {TagReq} from "./TagReq.tsx";
import type {CourseContentTranslate} from "./CourseContentTranslate.tsx";

export interface CourseTranslate {
    id: number;
    language: Language;
    titleShort: string;
    title: string;
    descriptionShort: string;
    description: string;
    descriptionLong: string;
    whatWillBeLearned: string[];
    skillsGained: TagReq[];
    topicsCovered: TagReq[];
    courseContents: CourseContentTranslate[];
}